+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @cat-factory/provider-s3
 
+## 0.2.65
+
+### Patch Changes
+
+- Updated dependencies [e0aab3f]
+  - @cat-factory/kernel@0.70.2
+
+## 0.2.64
+
+### Patch Changes
+
+- Updated dependencies [0d51638]
+  - @cat-factory/kernel@0.70.1
+
+## 0.2.63
+
+### Patch Changes
+
+- Updated dependencies [eb67d40]
+  - @cat-factory/kernel@0.70.0
+
 ## 0.2.62
 
 ### Patch Changes